+++ test/args.js
@@ -1,0 +1,3 @@
+var argv = require('yargs').argv;
+
+console.log(argv);